+++ app.js
@@ -1,4 +1,4 @@
-const cloningUtility = require('./cloningUtility').CloningUtility;
+const cloningUtility = require('./utilities/cloningUtility').CloningUtility;
 
 cloningUtility.performClone();
 
@@ -6,7 +6,11 @@ console.log('-------------')
 console.log('')
 console.log('')
 console.log('-------------')
-const PartnersProfile = require('./partnersProfile').PartnersProfile;
-const partnersProfileStreamer = new PartnersProfile('./data/partners.json');
+const path = require( "path" );
+
+var partnersProfileJSONPath = path.resolve("./data/partners.json");
+console.log('reading partners from: ' + partnersProfileJSONPath);
+const PartnersProfile = require('./coordinates/partnersProfile').PartnersProfile;
+const partnersProfileStreamer = new PartnersProfile(partnersProfileJSONPath);
 
 partnersProfileStreamer.debugPrintPartnerProfilesWithinHundredKilometersOrderByCompanyName();
