+++ coordinates/partnersProfile.js
@@ -103,6 +103,10 @@ class PartnersProfile {
 
         // Note: Not using profiles.filter, because 
         //       this.filterPartnerProfileWithinHundredKilometers calls other functions from inside.
+        if(!profiles || !profiles.length){
+            console.log('No partners profile found');
+            return;
+        }
 
         const results = [];
         profiles.forEach(profile => {
